+++ lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_button/pages/login_page.dart';
 import 'package:flutter_button/pages/intro_page.dart';
-import 'package:flutter_button/pages/signup.dart';
 import 'package:flutter_button/pages/map_page.dart';
 
 void main() {
@@ -18,8 +16,6 @@ class Main extends StatelessWidget {
        home: const IntroPage(),
         routes: {
           '/intropage': (context) => const IntroPage(),
-          '/loginpage': (context) => LoginPage(),
-          '/sign_up': (context) => SignUpPage(),
           '/mappage': (context) => const MapPage(),
         });
   }
